30/06/2022 Итоговый проект 12 спринта Добавленна поддержка запроса на нахождение кратчайшего маршрута между остановками.

+++ transport-catalogue/domain.h
@@ -53,7 +53,7 @@ struct BusStat {
 };
 
 struct RoutingSetting {
-  double wait_time = 0;
+  int wait_time = 0;
   double bus_velocity = 0;
   
 };
